requirements.txt
demoparser2==0.38.0
numpy==2.2.5
pandas==2.2.3
polars==1.29.0
pyarrow==20.0.0
pygame-ce==2.5.3
pygame_gui==0.6.13
python-dateutil==2.9.0.post0
python-i18n==0.3.9
pytz==2025.2
six==1.17.0
tqdm==4.67.1
tzdata==2025.2
